requirements.txt
google-genai
edge-tts
moviepy
pillow
requests
pandas
numpy
